1 => super admin 2 => create policy and gate for models 3 => update controllers and testes for modules 4 => create new test for all modules

+++ modules/Mshm/Category/Http/Controllers/CategoryController.php
@@ -4,6 +4,7 @@ namespace Mshm\Category\Http\Controllers;
 
 use App\Http\Controllers\Controller;
 use Mshm\Category\Http\Requests\CategoryRequest;
+use Mshm\Category\Models\Category;
 use Mshm\Category\Repositories\CategoryRepo;
 use Mshm\Category\Responses\AjaxResponses;
 
@@ -18,18 +19,24 @@ class CategoryController extends Controller
 
     public function index()
     {
+        /** @noinspection PhpUnhandledExceptionInspection */
+        $this->authorize('manage', Category::class);
         $categories = $this->repo->all();
         return view('Categories::index', compact('categories'));
     }
 
     public function store(CategoryRequest $request)
     {
+        /** @noinspection PhpUnhandledExceptionInspection */
+        $this->authorize('manage', Category::class);
         $this->repo->store($request);
         return back();
     }
 
     public function edit($categoryId)
     {
+        /** @noinspection PhpUnhandledExceptionInspection */
+        $this->authorize('manage', Category::class);
         $category = $this->repo->findById($categoryId);
         $categories = $this->repo->allExceptById($categoryId);
         return view('Categories::edit', compact('category', 'categories'));
@@ -37,12 +44,16 @@ class CategoryController extends Controller
 
     public function update($categoryId, CategoryRequest $request)
     {
+        /** @noinspection PhpUnhandledExceptionInspection */
+        $this->authorize('manage', Category::class);
         $this->repo->update($categoryId, $request);
         return back();
     }
 
     public function destroy($categoryId)
     {
+        /** @noinspection PhpUnhandledExceptionInspection */
+        $this->authorize('manage', Category::class);
         $this->repo->delete($categoryId);
         AjaxResponses::successResponse();
     }
